ci: generate more prefetch and cpu hz freeloader

+++ .github/build_multiple_freeloader.sh
@@ -3,8 +3,10 @@
 SOC=${SOC:-demosoc}
 CORE=${CORE:-ux900}
 BOOT_MODE=${BOOT_MODE:-sd}
-SELVAR=${SELVAR:-CPU_HZ}
-VARLIST=${VARLIST:-"16000000,100000000"}
+SELVAR0=${SELVAR0:-CPU_HZ}
+VARLIST0=${VARLIST0:-"16000000,100000000"}
+SELVAR1=${SELVAR1:-SPFL1DCTRL1}
+VARLIST1=${VARLIST1:-"0x0"}
 GITSHA=${GITSHA:-$(git describe --always)}
 MAKEOPTS=${MAKEOPTS:-""}
 DRYRUN=${DRYRUN:-0}
@@ -24,33 +26,41 @@ if [[ "$CI_JOB_ID" =~ ^[0-9]+$ ]] ; then
     mkdir -p $SYNCDIR
 fi
 
-for var in ${VARLIST//,/ }
+for var0 in ${VARLIST0//,/ }
 do
-    echo "Build freeloader for $MAKEOPTS $SELVAR=$var"
-    eval export $SELVAR=$var
-    frlname=freeloader_${GITSHA}_${SOC}_${CORE}_${BOOT_MODE}
-    if [ "$SELVAR" != "SOC" ] && [ "$SELVAR" != "CORE" ] && [ "$SELVAR" != "BOOT_MODE" ] ; then
-        frlname=${frlname}_${SELVAR,,}-${var}
-    fi
-    frldelf=work/${SOC}/${frlname}.elf
-    runcmd="make freeloader && cp -f work/${SOC}/freeloader/freeloader.elf ${frldelf}"
-    echo $runcmd
-    if [ "x$DRYRUN" == "x0" ] ; then
-        eval $runcmd
-    fi
-    unset $SELVAR
-    if [ "x$SYNCDIR" != "x" ] ; then
-        echo "Sync freeloader $frldelf to ${SYNCDIR}"
-        cp -f $frldelf $SYNCDIR/
-    fi
+    for var1 in ${VARLIST1//,/ }
+    do
+        echo "Build freeloader for $MAKEOPTS $SELVAR0=$var0 $SELVAR1=$var1"
+        eval export $SELVAR0=$var0 $SELVAR1=$var1
+        frlname=freeloader_${GITSHA}_${SOC}_${CORE}_${BOOT_MODE}
+        if [ "$SELVAR0" != "SOC" ] && [ "$SELVAR0" != "CORE" ] && [ "$SELVAR0" != "BOOT_MODE" ] ; then
+            frlname=${frlname}_${SELVAR0,,}-${var0}
+        fi
+        if [ "$SELVAR1" != "SOC" ] && [ "$SELVAR1" != "CORE" ] && [ "$SELVAR1" != "BOOT_MODE" ] ; then
+            frlname=${frlname}_${SELVAR1,,}-${var1}
+        fi
+        frldelf=work/${SOC}/${frlname}.elf
+        runcmd="make freeloader && cp -f work/${SOC}/freeloader/freeloader.elf ${frldelf}"
+        echo $runcmd
+        if [ "x$DRYRUN" == "x0" ] ; then
+            eval $runcmd
+        fi
+        unset $SELVAR0 $SELVAR1
+        if [ "x$SYNCDIR" != "x" ] ; then
+            echo "Sync freeloader $frldelf to ${SYNCDIR}"
+            cp -f $frldelf $SYNCDIR/
+        fi
+    done
 done
 
 if [ "x$SYNCDIR" != "x" ] ; then
     echo "Link latest to $SYNCDIR"
-    rm -f $SHARELOC/latest
-    pushd $SHARELOC
-    ln -s $GITSHA latest
-    popd
+    if [ "x${CI_COMMIT_BRANCH}" == "xdev_nuclei_next" ]
+        rm -f $SHARELOC/latest
+        pushd $SHARELOC
+        ln -s $GITSHA latest
+        popd
+    fi
 fi
 
 exit 0
